stock_market_crawler/requirements.txt
pymysql==1.1.1; python_version >= '3.7'
pandas==2.0.3; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
sqlalchemy==2.0.36; python_version >= '3.7'
